refactor: move MetadataPopover/v2 -> MetadataPopover
