+++ operate/client/src/App/ProcessInstance/TopPanel/MetadataPopover/Details/buildMetadata.ts
