+++ operate/client/src/App/ProcessInstance/TopPanel/MetadataPopover/Details/getExecutionDuration.tsx
