+++ operate/client/src/App/ProcessInstance/TopPanel/MetadataPopover/Incidents/multiIncidents.tsx
@@ -6,7 +6,7 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {Header} from '../../Header';
+import {Header} from '../Header';
 import {SummaryDataValue} from '../styled';
 
 type Props = {count: number; onButtonClick: () => void};
