+++ operate/client/src/App/ProcessInstance/TopPanel/MetadataPopover/Incidents/resolveIncidentErrorType.ts
