+++ operate/client/src/App/ProcessInstance/TopPanel/MetadataPopover/Incidents/singleIncident.tsx
@@ -9,7 +9,7 @@
 import {Stack} from '@carbon/react';
 import {Link} from 'modules/components/Link';
 import {Paths} from 'modules/Routes';
-import {Header} from '../../Header';
+import {Header} from '../Header';
 import type {Incident} from '@camunda/camunda-api-zod-schemas/8.8';
 import {SummaryDataKey, SummaryDataValue} from '../styled';
 import {resolveIncidentErrorType} from './resolveIncidentErrorType';
